LayoutTests/fast/js/const-expected.txt

This test checks that const declarations in JavaScript work and are readonly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL x should be RIGHT. Was WRONG.
FAIL y should be RIGHT. Was WRONG.
PASS successfullyParsed is true

TEST COMPLETE
